+++ Sources/OpenTelemetrySdk/Trace/SpanBuilderSDK.swift
@@ -44,7 +44,7 @@ public class SpanBuilderSdk: SpanBuilder {
     private var totalNumberOfLinksAdded: Int = 0
     private var parentType: ParentType = .currentSpan
 
-    private var startEpochNanos: Int = 0
+    private var startEpochNanos: Int64 = 0
 
     public init(spanName: String,
                 instrumentationLibraryInfo: InstrumentationLibraryInfo,
@@ -115,7 +115,7 @@ public class SpanBuilderSdk: SpanBuilder {
         return self
     }
 
-    public func setStartTimestamp(startTimestamp: Int) -> Self {
+    public func setStartTimestamp(startTimestamp: Int64) -> Self {
         startEpochNanos = startTimestamp
         return self
     }
